% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SPLITT__OrderedTreeStringNodes}
\alias{SPLITT__OrderedTreeStringNodes}
\alias{Rcpp_SPLITT__OrderedTreeStringNodes-class}
\title{Rcpp module for the \code{OrderedTreeStringNodes}-class}
\description{
Rcpp module for the \code{OrderedTreeStringNodes}-class
}
